renamed viewallbtn to navbar
